% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_preprocessing.R
\name{compute_split}
\alias{compute_split}
\title{Compute the splits of an image.}
\usage{
compute_split(
  n_col,
  n_row,
  col_size,
  row_size,
  col_overlap = 0,
  row_overlap = 0
)
}
\arguments{
\item{n_col, n_row}{Total number of columns and rows in the image.}

\item{col_size, row_size}{Number of columns and rows in each split.}

\item{col_overlap, }{row_overlap Columns and rows to add to each side of a
split.}
}
\value{
A data.frame.
}
\description{
Compute the pixel postions (columns and row) to split an image.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
