% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_preprocessing.R
\name{create_borders}
\alias{create_borders}
\title{Get mirrors of a raster's borders.}
\usage{
create_borders(r, col_overlap, row_overlap)
}
\arguments{
\item{r}{A terra's object.}

\item{col_overlap, }{row_overlap Number of columns and rows to add (repeat)
around the given raster.}
}
\value{
A list of terra's rasters.
}
\description{
Create rasters by mirroring raster's borders. These rasters
could be later mosaicked with the given raster.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
