% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_textures.R
\name{.image_to_rgba}
\alias{.image_to_rgba}
\title{Export the bands of an image}
\usage{
.image_to_rgba(in_file, out_dir)
}
\arguments{
\item{in_file}{Path to a file.}

\item{out_dir}{Path to a directory for storing the results.}
}
\value{
A vector with the paths to the resulting files.
}
\description{
Exprot the first 4 bands of an image-file to independent files. This
funciton assumes they correspond to red, green, blue, and near infra-red.
}
