% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_textures.R
\name{k_textures}
\alias{k_textures}
\title{Segment an image using the k-textures algorithm}
\usage{
k_textures(red, green = NULL, blue = NULL, nir = NULL, k, out_file)
}
\arguments{
\item{red, green, blue, nir}{Paths to raster images corresponding to the red,
green, blue, and near infrared bands of the spectrum. When red is a file
with more than 3 bands, this function extracts the first 4 bands.}

\item{k}{Number of segments.}

\item{out_file}{Path to the file resulting from the segmentation.}
}
\value{
out_file.
}
\description{
Segment a satellite image using the k-textures algorithm. This is an
implementation of the algorithm descibe in 'K-textures, a self-supervised
hard clustering deep learning algorithm for satellite image segmentation"
by Fabien et al which was inspired the by the k-means algorithm. K-textures
takes an image and a number of classes, then it trains a Convolutional
Neural Network which is used to separate the pixels in the given image
according to their texture.
}
