% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_textures.R
\name{k_textures}
\alias{k_textures}
\title{Segment an image using the k-textures algorithm}
\usage{
k_textures(
  r,
  k,
  out_file,
  col_size = 128,
  row_size = 128,
  col_overlap = 4,
  row_overlap = 4
)
}
\arguments{
\item{r}{A terra package object. A raster image.}

\item{k}{Number of segments.}

\item{out_file}{Path to the file resulting from the segmentation.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A terra's raster.
}\if{html}{\out{</div>}}
}
\description{
Segment a satellite image using the k-textures algorithm. This is an
implementation of the algorithm descibe in 'K-textures, a self-supervised
hard clustering deep learning algorithm for satellite image segmentation"
by Fabien et al which was inspired the by the k-means algorithm. K-textures
takes an image and a number of classes, then it trains a Convolutional
Neural Network which is used to separate the pixels in the given image
according to their texture.
}
